+++ group_14/src/kernel/src/include/terminal.h
@@ -1,5 +1,9 @@
 // Source: https://wiki.osdev.org/Bare_Bones Fetched: 15.03.2023
 
+#include <stdint.h>
+#include <stddef.h> 
+#include <cstdarg>
+
 #ifndef TERMINAL_H
 #define TERMINAL_H
 
@@ -33,66 +37,24 @@ static inline uint16_t vga_entry(unsigned char uc, uint8_t color)
 	return (uint16_t) uc | (uint16_t) color << 8;
 }
  
-size_t strlen(const char* str) 
-{
-	size_t len = 0;
-	while (str[len])
-		len++;
-	return len;
-}
+size_t strlen(const char* str);
  
 static const size_t VGA_WIDTH = 80;
 static const size_t VGA_HEIGHT = 25;
  
-size_t terminal_row;
-size_t terminal_column;
-uint8_t terminal_color;
-uint16_t* terminal_buffer;
+extern size_t terminal_row;
+extern size_t terminal_column;
+extern uint8_t terminal_color;
+extern uint16_t* terminal_buffer;
  
-void terminal_initialize(void) 
-{
-	terminal_row = 0;
-	terminal_column = 0;
-	terminal_color = vga_entry_color(VGA_COLOR_LIGHT_GREY, VGA_COLOR_BLACK);
-	terminal_buffer = (uint16_t*) 0xB8000;
-	for (size_t y = 0; y < VGA_HEIGHT; y++) {
-		for (size_t x = 0; x < VGA_WIDTH; x++) {
-			const size_t index = y * VGA_WIDTH + x;
-			terminal_buffer[index] = vga_entry(' ', terminal_color);
-		}
-	}
-}
+extern void terminal_initialize(void);
  
-void terminal_setcolor(uint8_t color) 
-{
-	terminal_color = color;
-}
+extern void terminal_setcolor(uint8_t color);
  
-void terminal_putentryat(char c, uint8_t color, size_t x, size_t y) 
-{
-	const size_t index = y * VGA_WIDTH + x;
-	terminal_buffer[index] = vga_entry(c, color);
-}
- 
-void terminal_putchar(char c) 
-{
-	terminal_putentryat(c, terminal_color, terminal_column, terminal_row);
-	if (++terminal_column == VGA_WIDTH) {
-		terminal_column = 0;
-		if (++terminal_row == VGA_HEIGHT)
-			terminal_row = 0;
-	}
-}
- 
-void terminal_write(const char* data, size_t size) 
-{
-	for (size_t i = 0; i < size; i++)
-		terminal_putchar(data[i]);
-}
- 
-void terminal_writestring(const char* data) 
-{
-	terminal_write(data, strlen(data));
-}
+extern void terminal_putentryat(char c, uint8_t color, size_t x, size_t y);
+extern void terminal_putchar(char c);
+extern void terminal_write(const char* data, size_t size);
+extern void terminal_writestring(const char* data);
 
+extern int printf(const char* __restrict__ format, ...);
 #endif //TERMINAL_H
