+++ group_14/src/kernel/src/memory/malloc.cpp
@@ -1,5 +1,4 @@
 #include "../include/memory.h"
-#include "../../../stdlib/include/system.h"
 
 #define MAX_PAGE_ALIGNED_ALLOCS 32
