+++ group_14/src/kernel/src/memory/memory.cpp
@@ -1,4 +1,3 @@
-#include "../../../stdlib/include/system.h"
 #include "../include/memory.h"
 
 static uint32_t* page_directory = 0;   // Define a pointer to the page directory and initialize it to zero
